requirements.txt
Django==2.1.3
djangorestframework==3.8.2
psycopg2-binary==2.7.5
django-cors-headers==2.4.0